+++ js/prettier.config.mjs
@@ -1,5 +1,5 @@
 const config = {
-    plugins: [],
+    plugins: ['prettier-plugin-organize-imports'],
     semi: true,
     printWidth: 140,
     tabWidth: 4,
