+++ js/src/sdk/base.toolset.spec.ts
@@ -2,8 +2,8 @@ import { describe, it, expect, beforeAll } from "@jest/globals";
 import { ComposioToolSet } from "./base.toolset";
 import { getTestConfig } from "../../config/getTestConfig";
 import { ActionExecutionResDto } from "./client";
-import { TPostProcessor } from "./common/action";
-import { TPreProcessor } from "./common/action";
+import { TPostProcessor } from "../types/base_toolset";
+import { TPreProcessor } from "../types/base_toolset";
 
 describe("ComposioToolSet class tests", () => {
   let toolset: ComposioToolSet;
