+++ js/src/types/base.d.ts
@@ -1,0 +1,3 @@
+export type Optional<T> = T | null;
+export type Dict<T> = { [key: string]: T };
+export type Sequence<T> = Array<T>;
